MacOS X support work in progress
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libusbhxcfe/trunk@441 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ sources/usb_hxcfloppyemulator.c
@@ -367,7 +367,7 @@ int ftdichiplistener(HXCFLOPPYEMULATOR* floppycontext,USBHXCFE * hw_context)
 	int we_ret;
 	int init_failed,ftdierror;
 	int i;
-	unsigned long RxBytes,TxBytes;
+	int RxBytes,TxBytes;
 	unsigned char * srambuffer;
 	unsigned char * srambuffer2;
 	unsigned char * input_buffer;
